refactor(rust): make generic join builder more dry (#5439)
