+++ polars/polars-lazy/polars-pipe/Cargo.toml
@@ -12,6 +12,7 @@ description = "Lazy query engine for the Polars DataFrame library"
 enum_dispatch = "0.3"
 hashbrown.workspace = true
 num.workspace = true
+polars-arrow = { version = "0.25.1", path = "../../polars-arrow", default-features = false }
 polars-core = { version = "0.25.1", path = "../../polars-core", features = ["lazy", "private", "zip_with", "random"], default-features = false }
 polars-io = { version = "0.25.1", path = "../../polars-io", default-features = false }
 polars-plan = { version = "0.25.1", path = "../polars-plan", default-features = false, features = ["compile"] }
