+++ polars/polars-lazy/polars-pipe/src/executors/sinks/joins/mod.rs
@@ -1,7 +1,8 @@
 #[cfg(feature = "cross_join")]
 mod cross;
+mod generic_build;
 mod inner;
 
 #[cfg(feature = "cross_join")]
 pub(crate) use cross::*;
-pub(crate) use inner::GenericBuild;
+pub(crate) use generic_build::GenericBuild;
